+++ src/apis/user/notification.py
@@ -12,7 +12,8 @@ from functions import *
 
 
 async def get_list(request: Request, response: Response, authorization: str = Header(None), \
-    page: Optional[int] = 1, page_size: Optional[int] = 10, query: Optional[str] = '', status: Optional[int] = None, \
+    page: Optional[int] = 1, page_size: Optional[int] = 10, after_notificationid: Optional[int] = None, \
+        query: Optional[str] = '', status: Optional[int] = None, \
         order_by: Optional[str] = "notificationid", order: Optional[str] = "desc"):
     """Returns a list of notification of the authorized user"""
     app = request.app
@@ -48,13 +49,17 @@ async def get_list(request: Request, response: Response, authorization: str = He
             order = "desc"
         elif order_by == "content":
             order = "asc"
-    order = order.upper()
 
     limit = ""
     if status == 0:
-        limit += "AND status = 0"
+        limit += "AND status = 0 "
     elif status == 1:
-        limit += "AND status = 1"
+        limit += "AND status = 1 "
+    if after_notificationid is not None:
+        if order == "asc":
+            limit += f"AND notificationid >= {after_notificationid} "
+        elif order == "desc":
+            limit += f"AND notificationid <= {after_notificationid} "
 
     await app.db.execute(dhrid, f"SELECT notificationid, content, timestamp, status FROM user_notification WHERE uid = {uid} {limit} AND LOWER(content) LIKE '%{query}%' ORDER BY {order_by} {order} LIMIT {max(page-1, 0) * page_size}, {page_size}")
     t = await app.db.fetchall(dhrid)